% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Allstats.R
\name{Allstats}
\alias{Allstats}
\title{Automatically processes T-test, U-test, Anova, Scheffe(Anova Post-Hoc), Krukal Wallis, Dunn-test(BH adjusted,(Kurkal Wallis Post-Hoc)) while allowing adjustment of FDR}
\usage{
Allstats(Data, Adjust_p_value = T, Adjust_method = "BH")
}
\arguments{
\item{Data}{csv file with Header as False First column with Sample Second column with Multilevel(Mixomics) so that it can be compatible with other multivariate statistics Third column with Group information. Rest of the columns are the metabolites to be tested.}

\item{Adjust_p_value}{Set True if FDR adjustments are to be made. If not set False}

\item{Adjust_method}{adjustment methods frequently used. "holm", "hochberg", "hommel", "bonferroni", "BH", "BY","fdr", "none"}
}
\value{
List including Result Matrix of p-values, converted datas.
}
\description{
Automatically processes T-test, U-test, Anova, Scheffe(Anova Post-Hoc), Krukal Wallis, Dunn-test(BH adjusted,(Kurkal Wallis Post-Hoc)) while allowing adjustment of FDR
}
\examples{
D<-as.data.frame(cbind(LETTERS,rep(1,26),c(rep("A",7),rep("B",6),rep("C",13)),runif(26),runif(26)))
Datas<-rbind(c("Sample","Multilevel","Group","some metabolite1","some metabolite2"),D)
Allstats(Datas)


}
