% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aroc_compute.R
\name{aroc_compute}
\alias{aroc_compute}
\title{compute.threshold.YI.AROC.sp function modified to retrieve sensitivity and specificity}
\usage{
aroc_compute(
  object,
  newdata,
  parallel = c("no", "multicore", "snow"),
  ncpus = 1,
  cl = NULL
)
}
\arguments{
\item{data}{https://rdrr.io/cran/ROCnReg/src/R/compute.threshold.YI.AROC.sp.R}
}
\value{
compute.threshold.YI.AROC.sp with sensitivity and specificity
}
\description{
compute.threshold.YI.AROC.sp function modified to retrieve sensitivity and specificity
}
