% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AS_barplot.R
\name{AS_barplot}
\alias{AS_barplot}
\title{Automatically save barplots for all metabolites including asterisks for significance.}
\usage{
AS_barplot(
  data,
  asterisk = "t_test",
  significant_variable_only = F,
  color = c("#FF3300", "#FF6600", "#FFCC00", "#99CC00", "#0066CC", "#660099"),
  legend_position = "none",
  order = NULL,
  tip_length = 0.01,
  label_size = 2.88,
  step_increase = 0.05,
  width = 0.3,
  fig_width = NA,
  fig_height = NA
)
}
\arguments{
\item{data}{List inheriting from Allstats}

\item{asterisk}{Choose asterisk to plot ("Dunn","Scheffe","u_test","t_test")}

\item{significant_variable_only}{show significant variable only (T,F)}

\item{color}{colors used for ggplots.}

\item{order}{order of the groups; order = c("A","B","C")}

\item{tip_length}{significance tip length}

\item{label_size}{significance label size}

\item{step_increase}{significance step increase}

\item{width}{bar width}

\item{fig_width}{figure size}

\item{fig_height}{figure size}
}
\value{
ggbarplot
}
\description{
Automatically save barplots for all metabolites including asterisks for significance.
}
\examples{
D<-as.data.frame(cbind(LETTERS,rep(1,26),c(rep("A",7),rep("B",6),rep("C",13)),runif(26),runif(26)))
Datas<-rbind(c("Sample","Multilevel","Group","some metabolite1","some metabolite2"),D)
Test<-Allstats(Datas)
AS_barplot(Test,asterisk = "Dunn",significant_variable_only = F,color = c("#FF3300","#FF6600","#FFCC00","#99CC00","#0066CC","#660099") )
}
