% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AS_boxplot.R
\name{AS_boxplot}
\alias{AS_boxplot}
\title{Automatically save boxplots for all metabolites including asterisks for significance.}
\usage{
AS_boxplot(
  data,
  asterisk = "t_test",
  significant_variable_only = F,
  color = c("#FF3300", "#FF6600", "#FFCC00", "#99CC00", "#0066CC", "#660099"),
  legend_position = "none",
  order = NULL,
  tip_length = 0.01,
  label_size = 2.88,
  step_increase = 0.05,
  width = 0.3,
  fig_width = NA,
  fig_height = NA,
  X_text = 10,
  Y_text = 12
)
}
\arguments{
\item{data}{List inheriting from Allstats}

\item{asterisk}{Choose asterisk to plot ("Dunn","Scheffe","u_test","t_test")}

\item{significant_variable_only}{show significant variable only (T,F)}

\item{color}{colors used for ggplots.}

\item{legend_position}{legend position "none","left","right","bottom","top"}

\item{order}{order of the groups; order = c("A","B","C")}

\item{tip_length}{significance tip length}

\item{label_size}{significance label size}

\item{step_increase}{significance step increase}

\item{width}{box width}

\item{fig_width}{figure size}

\item{fig_height}{figure size}

\item{X_text}{X axis text size}

\item{Y_text}{Y axis title size}
}
\value{
ggboxplot
}
\description{
Automatically save boxplots for all metabolites including asterisks for significance.
}
\examples{
data(Data)
Test<-Allstats(Data)
AS_boxplot(Test,asterisk = "Dunn",significant_variable_only = F,color = c("#FF3300","#FF6600","#FFCC00","#99CC00","#0066CC","#660099") )
}
