% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AS_dotplot.R
\name{AS_dotplot}
\alias{AS_dotplot}
\title{Automatically save dotplots for all metabolites including asterisks for significance.}
\usage{
AS_dotplot(
  data,
  asterisk = "t_test",
  significant_variable_only = F,
  color = c("#FF3300", "#FF6600", "#FFCC00", "#99CC00", "#0066CC", "#660099"),
  legend_position = "none",
  order = NULL,
  tip_length = 0.01,
  label_size = 2.88,
  step_increase = 0.05,
  size = NULL,
  fig_width = NA,
  fig_height = NA
)
}
\arguments{
\item{data}{List inheriting from Allstats}

\item{asterisk}{Choose asterisk to plot ("Dunn","Scheffe","u_test","t_test")}

\item{significant_variable_only}{show significant variable only (T,F)}

\item{color}{colors used for ggplots.}

\item{legend_position}{legend position "none","left","right","bottom","top"}

\item{order}{order of the groups; order = c("A","B","C")}

\item{tip_length}{significance tip length}

\item{label_size}{significance label size}

\item{step_increase}{significance step increase}

\item{size}{dot size}

\item{fig_width}{figure size}

\item{fig_height}{figure size}
}
\value{
ggdotplot
}
\description{
Automatically save dotplots for all metabolites including asterisks for significance.
}
\examples{
data(Data)
Test<-Allstats(Data)
AS_dotplot(Test,asterisk = "Dunn",significant_variable_only = F,color = c("#FF3300","#FF6600","#FFCC00","#99CC00","#0066CC","#660099") )
}
