% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classification-data.R
\docType{data}
\name{Classification}
\alias{Classification}
\title{Sample metabolomics Classification data}
\format{
An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.
}
\usage{
data(Classification)
}
\description{
Sample metabolomics Classification data
}
\examples{
data(Classification)
}
\keyword{datasets}
