% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D_tran.R
\name{D_tran}
\alias{D_tran}
\title{Scale_Transformation of data file}
\usage{
D_tran(data, param = "None", save = F)
}
\arguments{
\item{param}{"None","Auto","log10","Pareto"}

\item{save}{T,F}

\item{Data}{uploaded data}
}
\value{
scaled, transformed CSV file
}
\description{
Scale_Transformation of data file
}
\examples{
data(Data)
D_tran(Data,param = "Pareto")
}
