% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data-data.R
\docType{data}
\name{Data}
\alias{Data}
\title{Sample metabolomics data}
\format{
An object of class \code{data.frame} with 64 rows and 12 columns.
}
\usage{
data(Data)
}
\description{
Sample metabolomics data
}
\examples{
data(Data)
}
\keyword{datasets}
