% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_summary.R
\name{data_summary}
\alias{data_summary}
\title{modified summary function used for ggbarplot}
\usage{
data_summary(data, varname, groupnames)
}
\arguments{
\item{data}{http://www.sthda.com/english/wiki/ggplot2-error-bars-quick-start-guide-r-software-and-data-visualization}
}
\value{
summarized matrix with SEM and mean for Barplot
}
\description{
modified summary function used for ggbarplot
}
