% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Indiv_Perm.R
\name{Indiv_Perm}
\alias{Indiv_Perm}
\title{Computes PERMANOVA with vegan using specific group information}
\usage{
Indiv_Perm(Data, method = "bray")
}
\arguments{
\item{Data}{csv file with Header as False First column with Sample Second column with Multilevel(Mixomics) so that it can be compatible with other multivariate statistics Third column with Group information. Rest of the columns are the metabolites to be tested.}

\item{method}{Dissimilarity index c("manhattan", "euclidean", "canberra", "clark", "bray", "kulczynski", "jaccard", "gower", "altGower", "morisita", "horn", "mountford", "raup", "binomial", "chao", "cao", "mahalanobis", "chisq",chord")}
}
\value{
vegan::adonis2(Dist_Met~Group,method = "bray",by = NULL,data = x_y_coord_G)
}
\description{
Computes PERMANOVA with vegan using specific group information
}
\examples{
Df<-data(Data)
Indiv_Perm(Df)
}
