\docType{package}
\name{LMSstat-package}
\alias{LMSstat}
\alias{LMSstat-package}
\title{LMSstat: automation of statistical tests with an identical data input.}
\description{
Automation of statistical tests with an identical data input.
  The package Includes most Statistical tests used in metabolomics,
  aiming to reduce arduous work searching for packages and changing data input.

The package includes

* Simple Statistics :u-test, t-test, post hocs of Anova and Kruskal Wallis with FDR adjusted values

* Bar, Box, Dot plots with significance (u-test, t-test, post hocs of Anova and Kruskal Wallis)

* Scaling

* Heatmap

* Shapiro test

* PERMANOVA

* ROC and AROC

* NMDS

* PCA

* PCoA
}
\details{
\tabular{ll}{
Package: \tab LMSstat\cr
Type: \tab Package\cr
Version: \tab 1.0.8\cr
Date: \tab 2021-07-29\cr
License: \tab MIT License\cr
}
}
\author{
\strong{Maintainer}: Chang-Ho Kim \email{demizor3@gmail.com}

}
\keyword{internal}
