\name{NEWS}
\alias{NEWS}
\title{News for Package}

\section{Changes in version 1.0.1,(February 2021)}{
  \itemize{
    \item Package first made
    \item Allstats
    \item AS_barplot
    \item AS_boxplot
    \item AS_dotplot
    \item Indiv_Perm
    \item Multi_Perm
    \item roc_all
    \item Auxillary functions (aroc_compute, data summary)
  }
}
\section{Changes in version 1.0.2,(July 2021)}{
  \itemize{
    \item AS_barplot minor bug Fix
    \item AS_boxplot minor bug Fix
    \item AS_dotplot minor bug Fix
  }
}
\section{Changes in version 1.0.3,(July 2021)}{
  \itemize{
    \item AS_barplot Optimization for paper ready figures
    \item AS_boxplot Optimization for paper ready figures
    \item AS_dotplot Optimization for paper ready figures
  }
}
\section{Changes in version 1.0.4,(July 2021)}{
  \itemize{
  \item NMDS added
  }
}
\section{Changes in version 1.0.5,(July 2021)}{
  \itemize{
  \item PCA added
  }
}
\section{Changes in version 1.0.6,(July 2021)}{
  \itemize{
  \item Scaling added
  \item AS_heatmap added
  }
}
\section{Changes in version 1.0.7,(July 2021)}{
  \itemize{
  \item Shapiro test added
  }
}
\section{Changes in version 1.0.8,(July 2021)}{
  \itemize{
  \item PCoA added
  }
}
\section{Changes in version 1.0.9,(August 2021)}{
  \itemize{
  \item Volcano plot added
  }
}
\section{Changes in version 1.0.10,(August 2021)}{
  \itemize{
  \item Two factor analysis added
  }
}
