% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMDS.R
\name{NMDS}
\alias{NMDS}
\title{Automatically save NMDS plots for the selected dimension metrics}
\usage{
NMDS(
  Data,
  methods = c("euclidean", "bray"),
  color = c("#FF3300", "#660099", "#FFCC00", "#99CC00", "#0066CC", "#FF6600"),
  legend_position = "bottom",
  fig_width = 24,
  fig_height = 20
)
}
\arguments{
\item{Data}{}

\item{methods}{#methods<- c("manhattan", "euclidean", "canberra", "clark", "bray", "kulczynski", "jaccard", "gower", "altGower", "morisita", "horn",  "raup", "binomial", "mahalanobis", "chisq","chord") #can be any distance c\link{"manhattan", "euclidean", "canberra", "clark", "bray", "kulczynski", "jaccard", "gower", "altGower", "morisita", "horn", "mountford", "raup", "binomial", "chao", "cao", "mahalanobis", "chisq","chord"}}

\item{color}{colors used for ggplots.color=c("#FF3300","#FF6600","#FFCC00","#99CC00","#0066CC","#660099")}

\item{legend_position}{legend position "none","left","right","bottom","top"}

\item{fig_width}{figure size}

\item{fig_height}{figure size}
}
\value{
NMDS plot
}
\description{
Automatically save NMDS plots for the selected dimension metrics
}
\examples{
data(Data) 
NMDS(Data,methods = c("euclidian","manhattan","bray"))
}
