% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Norm_test.R
\name{Norm_test}
\alias{Norm_test}
\title{Normality test with Shaprio_Wilk_test}
\usage{
Norm_test(Data)
}
\arguments{
\item{Data}{input data file}
}
\value{
dataframe with shapiro wilk test
}
\description{
Normality test with Shaprio_Wilk_test
}
\examples{
Result<-Norm_test(Data)
write.csv(Result,"Normality_test_Result.csv")

}
