% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA.R
\name{PCA}
\alias{PCA}
\title{Automatically save PCA plot for the selected dimension metrics}
\usage{
PCA(
  Data,
  color = c("#FF3300", "#660099", "#FFCC00", "#99CC00", "#0066CC", "#FF6600"),
  legend_position = "none",
  fig_width = 24,
  fig_height = 20,
  components = c(1, 2),
  names = F,
  dotsize = 3,
  labsize = 3,
  ellipse = T
)
}
\arguments{
\item{Data}{uploaded data}

\item{color}{colors used for ggplots.color=c("#FF3300","#FF6600","#FFCC00","#99CC00","#0066CC","#660099")}

\item{legend_position}{legend position "none","left","right","bottom","top"}

\item{fig_width}{figure size}

\item{fig_height}{figure size}

\item{components}{PCA components c(1,3)}

\item{names}{whether the names are to be indicated T,F}

\item{dotsize}{dot size}

\item{labsize}{label size}

\item{ellipse}{T, F}
}
\value{
PCA plot
}
\description{
Automatically save PCA plot for the selected dimension metrics
}
\examples{
data(Data)
PCA(Data,components = c(1,2),legend_position = "none")
}
