% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCoA.R
\name{PCoA}
\alias{PCoA}
\title{Automatically save PCoA plots for the selected dimension metrics}
\usage{
PCoA(
  Data,
  methods = c("manhattan", "bray"),
  color = c("#FF3300", "#660099", "#FFCC00", "#99CC00", "#0066CC", "#FF6600"),
  legend_position = "none",
  components = c(1, 2),
  fig_width = 24,
  fig_height = 20,
  dotsize = 3,
  names = F,
  labsize = 3,
  ellipse = T
)
}
\arguments{
\item{Data}{uploaded data}

\item{methods}{c("manhattan", "euclidean", "canberra", "clark", "bray", "kulczynski", "jaccard", "gower", "altGower", "morisita", "horn",  "raup", "binomial", "mahalanobis", "chisq","chord")}

\item{color}{colors used for ggplots.color=c("#FF3300","#FF6600","#FFCC00","#99CC00","#0066CC","#660099")}

\item{legend_position}{legend position "none","left","right","bottom","top"}

\item{components}{PCA components c(1,3)}

\item{fig_width}{figure size}

\item{fig_height}{figure size}

\item{dotsize}{dot size}

\item{names}{whether the names are to be indicated T,F}

\item{labsize}{label size}

\item{ellipse}{T, F}
}
\value{
PCoA plot
}
\description{
Automatically save PCoA plots for the selected dimension metrics
}
\examples{
data(Data)
PCoA(Data,methods = c("euclidian","manhattan","bray"))

}
