% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_all.R
\name{roc_all}
\alias{roc_all}
\title{Computes AUC of AROC and ROC along with Specificity and Sensitivity of the Youden Index}
\usage{
roc_all(data, cov = T)
}
\arguments{
\item{data}{Healthy control as 0 and Disease control as 1. csv file with header as false. First column defining Sample information. Second column = Group, Third Column = Covariate ( Must be included even with random data to process the code). Rest of the columns are to be used as metabolites}

\item{cov}{T,F argument specifying whether covariate adjsuted ROC will be computed}
}
\value{
Either a list (adjusted + not adjusted) or a data matrix
}
\description{
Computes AUC of AROC and ROC along with Specificity and Sensitivity of the Youden Index
}
\examples{

D<-as.data.frame(cbind(paste0("sample_",1:26),c(rep(1,13),rep(0,13)),c(rep("2",7),rep("3",12),rep("4",7)),runif(26),runif(26)))
Datas<-rbind(c("Sample","Group","Covariate","some metabolite1","some metabolite2"),D)
roc_all(Datas)
}
