% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRH.R
\name{SRH}
\alias{SRH}
\title{Computes Scheirer-Ray-Hare Test}
\usage{
SRH(data, Adjust_p_value = T, Adjust_method = "BH")
}
\arguments{
\item{data}{csv file with Header as False First column with Sample Second column with Multilevel(Mixomics) so that it can be compatible with other multivariate statistics Third column with Group information. Rest of the columns are the metabolites to be tested.}

\item{Adjust_p_value}{Set True if FDR adjustments are to be made. If not set False}

\item{Adjust_method}{adjustment methods frequently used. "holm", "hochberg", "hommel", "bonferroni", "BH", "BY","fdr", "none"}
}
\value{
csv files including significant variables (Multilevel, Group, interaction) and a Venn diagram
}
\description{
Computes Scheirer-Ray-Hare Test
}
\examples{
data(Data)
Result<-SRH(Data)
}
