% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Volcano.R
\name{Volcano}
\alias{Volcano}
\title{Volcano plot for every group comparison}
\usage{
Volcano(
  data,
  asterisk = "t-test",
  reverse = F,
  fig_width = NA,
  FC_log = 2,
  pval_log = 10,
  fig_height = NA,
  dotsize = 2,
  fixed_limit = F,
  x_limit = c(-15, 15),
  y_limit = c(-2, 6),
  pval_intercept = 0.05,
  max_overlap = 20,
  FC_range = c(-1.5, 1.5),
  sig_label = T,
  color = c("blue", "black", "red")
)
}
\arguments{
\item{data}{data inheriting from Allstats}

\item{asterisk}{statistics inheriting from Allstats  # "t-test", "u-test", "Scheffe", "Dunn"}

\item{reverse}{reverse the direction of fold change}

\item{fig_width}{figure size}

\item{FC_log}{Fold change log transformation value}

\item{pval_log}{p_value log transformation value}

\item{fig_height}{figure size}

\item{dotsize}{dotsize}

\item{fixed_limit}{whether the limit should be fixed or not T, F}

\item{x_limit}{x axis limt c( -3,3)}

\item{y_limit}{y axis limit c(0, 100)}

\item{pval_intercept}{intercept for identification 0.05}

\item{max_overlap}{maximum overlap for labels}

\item{FC_range}{significant fold change range}

\item{sig_label}{label significant variables}

\item{color}{colors used for ggplots.color=c("#FF3300","#FF6600","#FFCC00","#99CC00","#0066CC","#660099")}
}
\value{
Volcano Plot
}
\description{
Volcano plot for every group comparison
}
\examples{
data(Data)
Test<-Allstats(Data)
Volcano(Test,asterisk = "t-test")
}
