% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMSstat-package.R
\docType{package}
\name{LMSstat-package}
\alias{LMSstat}
\alias{LMSstat-package}
\title{LMSstat: AUTOMATION OF STATISTICAL TESTS WITH AN IDENTICAL DATA INPUT}
\description{
Automation of statistical tests with an identical data input. The package Includes most Statistical tests used in metabolomics, aiming to reduce arduous work searching for packages and changing data input.
}
\author{
\strong{Maintainer}: Daehwan Kim \email{kdh5358@snu.ac.kr}

Authors:
\itemize{
  \item CHANG-HO KIM \email{demizor3@gmail.com}
}

}
\keyword{internal}
