% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multi_Perm.R
\name{Multi_Perm}
\alias{Multi_Perm}
\title{Computes PERMANOVA with vegan using multiple group information}
\usage{
Multi_Perm(Data, Classification, method = "bray")
}
\arguments{
\item{Data}{csv file with Header as False First column with Sample Second column with Multilevel(Mixomics) so that it can be compatible with other multivariate statistics Third column with Group information. Rest of the columns are the metabolites to be tested.}

\item{Classification}{csv file with Header as False ;First column with Sample; every other column can be used to indicate multiple classifiers}

\item{method}{Dissimilarity index c("manhattan", "euclidean", "canberra", "clark", "bray", "kulczynski", "jaccard", "gower", "altGower", "morisita", "horn", "mountford", "raup", "binomial", "chao", "cao", "mahalanobis", "chisq",chord")}
}
\value{
vegan::adonis2(Dist_Met~Group,method = "bray",by = NULL,data = x_y_coord_G)
}
\description{
Computes PERMANOVA with vegan using multiple group information
}
\examples{
D<-as.data.frame(cbind(LETTERS,rep(1,26),c(rep("A",7),rep("B",6),rep("C",13)),runif(26),runif(26)))
Datas<-rbind(c("Sample","Multilevel","Group","some metabolite1","some metabolite2"),D)
Datas_G<-rbind(c("Sample","some class0","some class1","some class2"),D)
Multi_Perm(Datas,Datas_G)
Note that the code recognizes each class as a factor.
}
