\name{NEWS}
\title{News for Package \pkg{LMSstat}}

\section{Changes in version 1.0.1,(February 2021)}{
  \itemize{
    \item Package first made
    \item Allstats
    \item AS_barplot
    \item AS_boxplot
    \item AS_dotplot
    \item Indiv_Perm
    \item Multi_Perm
    \item roc_all
    \item Auxillary functions (aroc_compute, data summary)
  }
}
\section{Changes in version 1.0.2,(July 2021)}{
  \itemize{
    \item AS_barplot minor bug Fix
    \item AS_boxplot minor bug Fix
    \item AS_dotplot minor bug Fix
  }
}
\section{Changes in version 1.0.3,(July 2021)}{
  \itemize{
    \item AS_barplot Optimization for paper ready figures
    \item AS_boxplot Optimization for paper ready figures
    \item AS_dotplot Optimization for paper ready figures
  }
}
\section{Changes in version 1.0.4,(July 2021)}{
  \itemize{
  \item NMDS added
  }
}
